Added entry to readme and made updates requested by Stephan
